+++ CHANGES.rst
@@ -1,3 +1,20 @@
+Changes in MatrixKit in 0.12.7 (2020-05-xx)
+=========================================
+
+Improvements:
+ * Upgrade MatrixSDK version ([v0.16.6](https://github.com/matrix-org/matrix-ios-sdk/releases/tag/v0.16.6)).
+ * DTCoreText: Update DTCoreText dependency to 1.6.23 minimum to be sure to not reference UIWebView.
+ * MXKCountryPickerViewController: Replace deprecated UISearchDisplayController by UISearchViewController.
+ * MXKLanguagePickerViewController: Replace deprecated UISearchDisplayController by UISearchViewController.
+ * MXKAppSettings: Add an option to hide un-decryptable events before joining the room.
+ * MXKRoomDataSource: Hide un-decryptable messages that were sent while the user was not in the room if needed.
+
+Bug fix:
+ * MXKRoomDataSource: Wait for store data ready when finalizing initialization on data source (vector-im/riot-ios/issues/3159).
+ * MXKLanguagePickerViewController: Fix selected cell reuse issue.
+ * MXKRoomDataSource: Wait for initial event existence if provided (vector-im/riot-ios/issues/3290).
+ * MXKRoomDataSource: Convert one-time observers to block variables to avoid releasing (vector-im/riot-ios/issues/3337).
+
 Changes in MatrixKit in 0.12.6 (2020-05-18)
 =========================================
 
@@ -18,6 +35,7 @@ Improvements:
 
 Bug fix:
  * Replace UIWebView with WKWebView (PR #663).
+ * Fix range of allowed surrogate emoji characters to 0x1d000-0x1f9ff.
 
 Changes in MatrixKit in 0.12.3 (2020-05-07)
 =========================================
@@ -1402,4 +1420,3 @@ MatrixKit contains the following reusable UI components:
  * MXKRecentListViewController
  * MXKRoomMemberListViewController
  
-